LayoutTests/ietestcenter/Javascript/15.4.4.15-8-9-expected.txt

15.4.4.15-8-9 (TestCases/chapter15/15.4/15.4.4/15.4.4.15/15.4.4.15-8-9.js) Array.prototype.lastIndexOf must return correct index (Sparse Array)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
